refactor(llamabot)🔧: Refactor logging and database path setup
- Centralize the find_or_set_db_path function in utils.py, removing redundancy.
- Implement dynamic log level setting based on environment variable.
- Ensure the .llamabot directory exists at the user's home for database logging.
